feat: deposit through shares wrapper
