+++ packages/sdk/src/Asset.ts
@@ -194,59 +194,3 @@ export async function getCanonicalValue(
 
   return result;
 }
-
-export async function getCanonicalAssetValue(
-  client: PublicClient,
-  args: Viem.ContractCallParameters<{
-    valueInterpreter: Address;
-    baseAsset: Address;
-    amount: bigint;
-    quoteAsset: Address;
-  }>,
-) {
-  try {
-    const { result } = await Viem.simulateContract(client, args, {
-      abi: Abis.IValueInterpreter,
-      functionName: "calcCanonicalAssetValue",
-      address: args.valueInterpreter,
-      args: [args.baseAsset, args.amount, args.quoteAsset],
-    });
-
-    return result;
-  } catch (error) {
-    // TODO: More selectively catch this error here.
-    if (error instanceof ContractFunctionExecutionError) {
-      return undefined;
-    }
-
-    throw error;
-  }
-}
-
-export async function calcCanonicalAssetsTotalValue(
-  client: PublicClient,
-  args: Viem.ContractCallParameters<{
-    valueInterpreter: Address;
-    baseAssets: Address[];
-    amounts: bigint[];
-    quoteAsset: Address;
-  }>,
-) {
-  try {
-    const { result } = await Viem.simulateContract(client, args, {
-      abi: Abis.IValueInterpreter,
-      functionName: "calcCanonicalAssetsTotalValue",
-      address: args.valueInterpreter,
-      args: [args.baseAssets, args.amounts, args.quoteAsset],
-    });
-
-    return result;
-  } catch (error) {
-    // TODO: More selectively catch this error here.
-    if (error instanceof ContractFunctionExecutionError) {
-      return undefined;
-    }
-
-    throw error;
-  }
-}
