+++ packages/sdk/src/Performance.ts
@@ -1,5 +1,5 @@
 import * as Abis from "@enzymefinance/abis";
-import type { Address, PublicClient } from "viem";
+import { type Address, ContractFunctionExecutionError, type PublicClient } from "viem";
 import { Viem } from "./Utils.js";
 
 export async function getNav(
@@ -111,4 +111,60 @@ export async function getSharePriceInAsset(
   });
 
   return result;
+}
+
+export async function getCanonicalAssetValue(
+  client: PublicClient,
+  args: Viem.ContractCallParameters<{
+    valueInterpreter: Address;
+    baseAsset: Address;
+    amount: bigint;
+    quoteAsset: Address;
+  }>,
+) {
+  try {
+    const { result } = await Viem.simulateContract(client, args, {
+      abi: Abis.IValueInterpreter,
+      functionName: "calcCanonicalAssetValue",
+      address: args.valueInterpreter,
+      args: [args.baseAsset, args.amount, args.quoteAsset],
+    });
+
+    return result;
+  } catch (error) {
+    // TODO: More selectively catch this error here.
+    if (error instanceof ContractFunctionExecutionError) {
+      return undefined;
+    }
+
+    throw error;
+  }
+}
+
+export async function calcCanonicalAssetsTotalValue(
+  client: PublicClient,
+  args: Viem.ContractCallParameters<{
+    valueInterpreter: Address;
+    baseAssets: Address[];
+    amounts: bigint[];
+    quoteAsset: Address;
+  }>,
+) {
+  try {
+    const { result } = await Viem.simulateContract(client, args, {
+      abi: Abis.IValueInterpreter,
+      functionName: "calcCanonicalAssetsTotalValue",
+      address: args.valueInterpreter,
+      args: [args.baseAssets, args.amounts, args.quoteAsset],
+    });
+
+    return result;
+  } catch (error) {
+    // TODO: More selectively catch this error here.
+    if (error instanceof ContractFunctionExecutionError) {
+      return undefined;
+    }
+
+    throw error;
+  }
 }
